refactor(quote): move the services into a config folder
