+++ src/features/quote/services/config/quote-config-core.service.spec.ts
@@ -1,6 +1,6 @@
 import { QuoteConfigCoreService } from './quote-config-core.service';
-import { ServiceNameEnum } from '../../../enums/service-name.enum';
-import { CoreEventService } from '../../core/services/core-event.service';
+import { ServiceNameEnum } from '../../../../enums/service-name.enum';
+import { CoreEventService } from '../../../core/services/core-event.service';
 
 describe(`QuoteConfigCoreService`, (): void => {
   let service: QuoteConfigCoreService;
