+++ src/features/quote/services/config/quote-config.service.ts
@@ -1,7 +1,7 @@
 import { QuoteConfigCoreService } from './quote-config-core.service';
-import { AbstractService } from '../../../classes/services/abstract.service';
-import { ServiceNameEnum } from '../../../enums/service-name.enum';
-import { IQuoteConfig } from '../interfaces/quote-config';
+import { AbstractService } from '../../../../classes/services/abstract.service';
+import { ServiceNameEnum } from '../../../../enums/service-name.enum';
+import { IQuoteConfig } from '../../interfaces/quote-config';
 import _ from 'lodash';
 
 export class QuoteConfigService extends AbstractService {
